no big changes

+++ src/Radication/Radication.cs
@@ -1,5 +1,4 @@
 using System;
-
 namespace Radication
 {
     public class Radicacion
@@ -11,7 +10,6 @@ namespace Radication
             {
                 throw new ArgumentException("0 no puede ser una potencia para una raiz");
             }
-
             this.potency = potency;
         }
 
@@ -21,8 +19,9 @@ namespace Radication
             {
                 throw new ArgumentException("No puede tener como potencia u nnumero par y sacarle la raiz a un numero positivo");
             }
-
-            double answer = Math.Pow(number, 1 / potency);
+            
+            double answer = Math.Pow(number, (double) 1 / potency);
+            
 
             return answer;
         }
